+++ src/main/java/com/tagnumelite/chickens/datagen/ModDataGenerators.java
@@ -1,6 +1,8 @@
 package com.tagnumelite.chickens.datagen;
 
 import com.tagnumelite.chickens.Chickens;
+import com.tagnumelite.chickens.datagen.chickens.VanillaChickenProvider;
+import com.tagnumelite.chickens.datagen.fluid_eggs.VanillaFluidEggProvider;
 import net.minecraft.data.DataGenerator;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.data.event.GatherDataEvent;
@@ -14,6 +16,8 @@ public class ModDataGenerators {
         DataGenerator generator = event.getGenerator();
         ExistingFileHelper existingFileHelper = event.getExistingFileHelper();
 
+        generator.addProvider(event.includeServer(), new VanillaChickenProvider(generator, existingFileHelper));
+        generator.addProvider(event.includeServer(), new VanillaFluidEggProvider(generator, existingFileHelper));
         generator.addProvider(event.includeServer(), new ModEntityTagProvider(generator, existingFileHelper));
 
         generator.addProvider(event.includeClient(), new ModBlockStateProvider(generator, existingFileHelper));